Implemented BrightnessBuffer
BrightnessBuffer will be used by a Renderer to draw pixels
from parametric surfaces onto a buffer. This buffer implements
ToString so that it can be printed.

The levels array determines which characters are used for
ranges in brightness. When editing the array, make sure
the minBrightness values are in descending order